Make categorical re-ordering quieter
